XPCSPI.h build fix for "[visionOS] Add SPI module for LinearMediaKit" rdar://127327242 https://bugs.webkit.org/show_bug.cgi?id=273527
Reviewed by NOBODY (OOPS!).

In https://commits.webkit.org/278232@main, XPCSPI.h was added to the
include graph of WebKitSwift on visionOS. This header imports
<AppServerSupport/OSLaunchdJob.h>, which is fine when included
textually, but when imported as a module it pulls in AppServerSupport.h.
That header contains an error diagnostic which fails the build (because
the framework is only supposed to be used by XPC servers).

Fix by moving the OSLaunchdJob.h import out of XPCSPI.h and into the two
places where it's actually used. This will likely be a problem again if
webpushd starts building with modules enabled, but it will be isolated
to the target which actually uses this class.

* Source/WTF/wtf/spi/darwin/XPCSPI.h:
* Source/WebKit/webpushd/webpushtool/WebPushToolMain.mm:
* Tools/TestWebKitAPI/cocoa/DaemonTestUtilities.mm: