rot13 API: DO NOT MERGE.
This is an attempt at adding a pointless web API just so I can learn about how
it's done in WebKit.

This API calls through to the UI process to request some processing. It doesn't
yet do anything and this is incomplete.

The API can be exercised from devtools using:

  window.navigator.rot13("Some text")