Abstract message receipt out of WebPageProxy (do not merge).
This is an experimental hacky change to test what happens if we split
WebPageProxy into two:
- the parts responsible for receiving messages from the web process
- all the rest

The first part is obviously very security-sensitive and is a key layer we'd
need to convert into a memory safe language in future if we ever wanted to
handle receipt of _any_ messages from the WebProcess in a memory-safe
language end-to-end.

This is an experimental change full of hacks, and has been used to
gather some data, mostly about the calls from the new "message handler"
back into the rest of the system.

The browser launches but has not been remotely tested. It's highly
likely that I made errors in handling dozens of the messages - this
is a throwaway protoype. Do not merge.