Swift code in core WebKit [do not merge]
This is an experimental change to see what goes wrong when we include Swift code within core WebKit.

This attempts to ensure that we have warnings treated as errors, and Strictly Safe Swift enabled.

It also ensures that C++ has access to the correct function prototypes to call. This slightly
conflicted with "warnings as errors" since normally this emits warnings about
a lack of availability annotations. This commit aims to include command line flags to suppress
such warnings.