Resolve TAPI problems with WebGPU swift. https://bugs.webkit.org/show_bug.cgi?id=291363
Reviewed by NOBODY (OOPS!).

It appears that the Swift build process generates some synthetic symbols which
end up being exported from the dylib, but are not described in the TBD file,
and this causes the TAPI tool to fail the build. This change explicitly asks
the linker to cease to include them in the dylib.

* Source/WebGPU/Configurations/WebGPU.xcconfig: