Reinstate RefPtr errors.
This commit re-enables some code which causes Swift/C++ interop to fail due to clang modularization problems.
It seems that the problem is that _is_constructible is unable to figure out whether RefPtr is constructible
due to having access only to a partial definition without default template parameters. This seems
to be because the full definition has ended up in the clang module WebKit_Private.WKRetainPtr.

* Source/WebKit/UIProcess/WebBackForwardList.h: