Add WTF and bmalloc modulemaps https://bugs.webkit.org/show_bug.cgi?id=292791 rdar://151023888
Reviewed by NOBODY (OOPS!).

This commit adds modulemap files for WTF and bmalloc. These are necessary
because Swift/C++ interop interprets C++ headers using the "-fmodules" option,
interpreting the C++ codebase as clang modules.

The approach taken for the two modulemaps is slightly different. bmalloc
defines a single monolithic module, the simplest approach. For WTF we choose
a more complex approach, as previously followed by the WebKit_Internal
module, where the module is divided into many small submodules. This
approach is chosen because Swift/C++ interoperability otherwise tries
to interpret the whole module, and that results in errors such as:

  /usr/include/c++/v1/__memory/construct_at.h:47:58: note: candidate template ignored: substitution failure [with _Tp = WTF::TextBreakIterator, _Args = <const WTF::TextBreakIterator &>]: call to deleted constructor of 'WTF::TextBreakIterator'
_LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX20 _Tp* __construct_at(_Tp* __location, _Args&&... __args) {

This is presumed to be rdar://148760413 or related overinstantiation
of templates.  We'll likely need to solve those in future, but meanwhile
splitting the module into lots of submodules limits the blast radius of
such issues and allows us to make progress with Swift.

Notes on some specific parts of the WTF modulemap:
* HashMap and RobinHoodHashTable.h are interdependent so need to be
  colocated in a module.
* A few headers (e.g. Platform) are used in contexts without C++20 support,
  unlike other parts of WTF.
* MetadataSPI.h is included textually because its dependencies don't
  always seem to be present in the way clang modules expect.

We do not export the modulemaps to the build output on pre-OS X 15;
doing so confuses the build of PALSwift.

This also defines UCHAR_TYPE within PALSwift. This is necessary because
previously UCHAR_TYPE was defined by a prefix header; prefix headers
are not applied to modules.

This PR is the second attempt at landing modulemaps for WTF and bmalloc.
The last time didn't work because it turned out that we needed to resolve
three issues first, to unbreak downstream consumers:

* The UCHAR_TYPE issue mentioned above.
* Downstream consumers of WebKit were previously using some WTF definitions
  without including WTF headers. That ceases to work when WTF becomes a
  module; we need to land those fixes before landing these modulemaps.
* #45625 resolves an issue with AppServerSupport which starts to appear
  once a modulemap is in place.

* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/module.modulemap: Added.
* Source/bmalloc/bmalloc.xcodeproj/project.pbxproj:
* Source/bmalloc/module.modulemap: Added.