Add WTF and bmalloc modulemaps https://bugs.webkit.org/show_bug.cgi?id=292791 rdar://151023888
Reviewed by NOBODY (OOPS!).

This commit adds modulemap files for WTF and bmalloc. These are necessary
because Swift/C++ interop interprets C++ headers using the "-fmodules" option,
interpreting the C++ codebase as clang modules.

The approach taken for the two modulemaps is slightly different. bmalloc
defines a single monolithic module, the simplest approach. For WTF we choose
a more complex approach, as previously followed by the WebKit_Internal
module, where the module is divided into many small submodules. This
approach is chosen because Swift/C++ interoperability otherwise tries
to interpret the whole module, and that results in errors such as:

  /usr/include/c++/v1/__memory/construct_at.h:47:58: note: candidate template ignored: substitution failure [with _Tp = WTF::TextBreakIterator, _Args = <const WTF::TextBreakIterator &>]: call to deleted constructor of 'WTF::TextBreakIterator'
_LIBCPP_HIDE_FROM_ABI _LIBCPP_CONSTEXPR_SINCE_CXX20 _Tp* __construct_at(_Tp* __location, _Args&&... __args) {

This is presumed to be rdar://148760413 or related overinstantiation
of templates. We'll likely need to solve those in future, but meanwhile
splitting the module into lots of submodules limits the blast radius of
such issues and allows us to make progress with Swift.

The modulemaps themselves are quite repetitive and need to be kept
up to date with the addition or removal of header files. We therefore
don't include them in the code directly, but generate them using a
script and a TOML file that describes the quirks of given header files
(for example, whether they're interdependent or can be used in plain
C contexts). See the TOML files in this PR and their comments for
a description of these quirks. Thanks to Elliott Williams for writing
the script.

The script converts a tapi filelist-style header listing
into a modulemap, supporting both monolithic and submodule layouts.
This process works well for libraries like bmalloc and WTF which have no
umbrella header, and which have no concept of public vs private API.
(We could potentially have instead used an 'umbrella directory' for
bmalloc, but not for WTF as we require the submodules for the reasons
described above).

We do not export the modulemaps to the build output on pre-OS X 15;
doing so confuses the build of PALSwift.

This also defines UCHAR_TYPE within PALSwift. This is necessary because
previously UCHAR_TYPE was defined by a prefix header; prefix headers
are not applied to modules.

This PR is the second attempt at landing modulemaps for WTF and bmalloc.
The last time didn't work because it turned out that we needed to resolve
three issues first, to unbreak downstream consumers:

* The UCHAR_TYPE issue mentioned above.
* Downstream consumers of WebKit were previously using some WTF definitions
  without including WTF headers. That ceases to work when WTF becomes a
  module; we need to land those fixes before landing these modulemaps.
* #45625 resolves an issue with AppServerSupport which starts to appear
  once a modulemap is in place.

* Source/WTF/Configurations/WTF.xcconfig:
* Source/WTF/Configurations/modulemap.toml: Added.
* Source/WTF/Scripts/modulemap-from-tapi-filelist.py: Added.
* Source/WTF/WTF.xcodeproj/project.pbxproj:
* Source/WTF/wtf/Compiler.h:
* Source/WTF/wtf/SIMDUTF.cpp: simdutf_impl.cpp.h should not be treated
  as a library header, and it confuses the generator. It's only used to
  compile this file, so change it to a local quote-include.
* Source/WebCore/PAL/Configurations/PAL.xcconfig:
* Source/WebCore/PAL/pal/spi/cg/CoreGraphicsSPI.h:
* Source/WebKitLegacy/ios/Misc/WebUIKitSupport.h:
* Source/bmalloc/Configurations/bmalloc.xcconfig:
* Source/bmalloc/Configurations/modulemap.toml: Added.
* Source/bmalloc/bmalloc.xcodeproj/project.pbxproj:
* Source/bmalloc/libpas/src/libpas/pas_darwin_spi.h: