Add script to workaround upstream problem using VFS.
This adds a script which constructs a VFS overlay to work around a problem in a private framework.
This is clearly not OK for submission and we'll need to get the upstream issue fixed.
The need for this overlay will prevent this branch from building unless the overlay is created
which means this branch won't currently work against the external SDK.