PAL inherit OTHER_SWIFT_FLAGS - DO NOT MERGE rdar://164560176
This also adds 'unsafe' arbitrarily to all the places it's needed, without
assessing actual safety - do not merge.

Even if we felt this was OK, we can't really merge this as-is because this code
needs to compile with Swift 5.9 which does not understand the 'unsafe' keyword.
So we'd need to add #if swift(>6.0)/#else/#endif around many lines, leading to
lots of duplication.