Ref count various WebKit types within Swift https://bugs.webkit.org/show_bug.cgi?id=302644 rdar://164888628
Reviewed by NOBODY (OOPS!).

This informs Swift that some C++ types should be treated as Swift reference
types (i.e. reference-counted) rather than value types. It teaches Swift
how to manipulate the reference counts for these types.

This is a temporary solution. In due course, we'll be able to inform
Swift that the base WTF::RefCounted type is the ancestor of all reference
counted types, and it won't be necessary to make these changes to
individual classes.

That's enabled by either of these changes:
rdar://131065240 - allow templated types to be ref counted
rdar://160696723 - allow method calls for ref and deref
We can't currently use this functionality in WebKit because we need
to support older compilers, and also due to a compiler
bug rdar://162195228. So for now, we will annotate just those
specific types which are required for upcoming CoreIPC message handler
work.