Introduce WTF::Cell https://bugs.webkit.org/show_bug.cgi?id=303175 rdar://165481024
Reviewed by NOBODY (OOPS!).

This introduces a new API, WTF::Cell, the purpose of which is to attach a
reference count to any type, potentially even move-only types.

This is needed for Swift/C++ interoperability because rdar://162361370
prevents us from passing move-only types from C++ to Swift. We'll need
to do that in the near future for WTF::Function and
WTF::CompletionHandler as we start to handle IPC message reception in
Swift code.

Fortunately, we already had such a type - as an implementation detail
of WTF::Box. Extract it from there to make it a standalone type.
Because this is the existing type, it retains the thread-safe
reference counting semantics.

Alternatives considered:
* Having a ref counted type for each thing we want to pass from C++
  to Swift, e.g. WTF::RefCounted::Function and
  WTF::RefCounted::CompletionHandler. Having a single Cell wrapper
  is obviously simpler.
* Naming this type SwiftCell, or similar. This has the advantage that
  it would be more obvious to remove such usages when
  rdar://162361370 is fixed, but it was determined that the
  semantics of this type are not Swift-specific, so it should be
  named something general.
* Make Box simply:
    template<typename T>
    using Box = RefPtr<Cell<T>>.
  This turned out to be impractical as there are many usages of
  Box's operator-> and operator*, especially in WTF's locking code
  and its customers. It is definitely worth keeping Box and its
  overloaded operators to keep that code as simple as possible.