Allow const calls to CompletionHandler from Swift https://bugs.webkit.org/show_bug.cgi?id=303180 rdar://165483404
Reviewed by NOBODY (OOPS!).

Upcoming work to use Swift for CoreIPC message recievers will require us to
pass WTF::CompletionHandlers across the language boundary from C++ to Swift.
These are move-only types, and thus are not directly supported by current
Swift/C++ interop (this is rdar://162361370). Instead, we'll need to wrap them
in a WTF::Cell (being introduced in
https://github.com/WebKit/WebKit/pull/54517) or some similar reference-counted
container.

We can then use two Swift/C++ interop features to call methods on the CompletionHandler:
* .pointee allows access to the operator* of the Cell;
* thing() calls the operator() of the thing.
So, we can do box.pointee(). Almost. The problem is the 'const'ness, because
.pointee only allows const access to the contents.

In the case of CompletionHandler, the object is updated to ensure that the
handler is called exactly once, so the current operator() is not const. In
Swift, we add a const version to allow Box<CompletionHandler> to be used across
the language boundary like this. Although it's obviously slightly messy
to cast away the 'const' like this, the pattern is in keeping with the way the
same type is used in C++, and it should be tempoary pending a fix for
rdar://162361370.