IPC can call directly into Swift https://bugs.webkit.org/show_bug.cgi?id=302950 rdar://165213339
Reviewed by NOBODY (OOPS!).

This is a draft change seeking early review of an approach to handle IPC
messages directly in Swift code.

This is going to be a long commit message, dear reader. Perhaps get a
cup of tea and some jaffa cakes.

GOALS:

The goals of this approach are:
* We do not duplicate all the message decoding code. The existing C++
  autogenerated decoding code should be used to call into Swift handler
  functions, by means of Swift/C++ interop.
* Messages can be defined, as now, within .in files, and each message calls
  directly into Swift handler functions, with no need for manually-written
  intermediate C++ shim functions for each message.

The key part is there's no need for manually-written C++ code for EVERY
MESSAGE, but there IS a need for manually-written code to manage the lifetime
of the message handler types.

USER EXPERIENCE:

To make a Swift message handler, you must:

* Mark your messages.in with a new attribute, SwiftReceiver.
* Create a Swift class with suitable functions for each message. It does
  not need any particular inheritance or conformance to any particular
  protocol. It needs to be named the same as the type named in the
  messages.in, just as if you were using C++.
* Within that class, create a new message _forwarder_ class
  (this is explained below) using code like this:
     WebKit.WebThingyMessageForwarder.create(target: self);
  (the implementation of that 'create' function is autogenerated; you
  don't have to do that)
* Register _that_ object to receive messages, rather than the Swift
  class itself.
* Keep that forwarder object referenced until you don't need to
  receive messages any more.

All of this is explained below, but users don't really need to know
anything more than the above.

OBJECTS INVOLVED:

Somewhere, we have a Swift type, which actually handles the messsages.
Let's assume this is called WebThingy. This is a Swift class, thus
subject to reference counting. Note it's a pure Swift type, not some
hybrid C++/Swift type.

It contains methods which handle messages. Examples:

 @_expose(Cxx)
 internal func webThingyUpdateThingy(connection: IPC.Connection, frameState: RefFrameState)

We want the autogenerated code to call such methods.

However, our code generation currently generates C++ member functions
such as ::didReceiveMessage. As WebThingy is a Swift type, we can't
possibly attach C++ member functions to this type.

As such, we need an extra type. We're going to autogenerate this within
our Python code generation. We will call it WebThingyMessageForwarder.
It's a pure C++ type. It exists to accommodate the ::didReceiveMessage
(and similar) member functions. It does not have member functions for
each individual message; it's entirely autogenerated. It knows how to
forward actual individual method calls onto the real WebThingy (more
details on that later).

The code which registers WebThingy as a message receiver can no longer
register the WebThingy itself. Instead, it needs to register the
WebThingyMessageForwarder.

OWNERSHIP:

The ownership here is a little fiddly.

The WebThingy OWNS the WebThingyMessageForwarder, via a Ref<T>.
This is necessary because the IPC message infrastructure only retains
weak references to message receiver objects, so if it's not owned
elsewhere, it would be deallocated and no messages would be
received.

But, of course, the WebThingyMessageForwarder needs to call methods
on the Swift WebThingy itself. To avoid a reference loop, the
WebThingyMessageForwarder has a weak reference to WebThingy.
Unfortunately, Swift/C++ interop does not yet have built-in
facilities for C++ to keep a weak Swift reference (this is
rdar://165068975) so an extra WebThingyWeakRef class needs to be
created to fulfil this purpose. But it's essentially just a weak
reference.

When the WebThingyMessageForwarder wants to call a WebThingy
method, it first calls WebThingyMessageForwarder::getTarget().
This will briefly turn the weak reference into a strong reference
while the method call occurs.

Only some of this is visible in this PR. Bits which are visible:
* the definition of the *MessageForwarder type
* definition of the WebThingyWeakRef type
* definition of getTarget()
* calling getTarget()
* MessageForwarder 'create' function and all the dancing around weak
  references

Bits which are not visible:
* creation and registration of the MessageForwarder

DISPATCH:

So, when C++ calls WebThingyMessageForwarder::didReceiveMessage,
that decodes the message body in the normal way, and then
all this happens:
* didReceiveMessage calls WebThingyMessageForwarder::getTarget
* that converts the weak reference briefly into a strong
  reference
* that strong reference is used to call the actual message
  dispatch function on the Swift WebThingy.
* Swift handles the message.

MESSAGES.IN CHANGES:

There's only one change to messages.in file, which is an optional
SwiftReceiver attribute. It's a Boolean. If that's enabled, all
the above stuff happens.

PASSING IPC CONNECTIONS:

Some message handlers like to receive an IPC::Connection. Currently,
they're passed as a C++ reference, which is not compatible with
Swift/C++ interop. So, if these functions are Swift, we need to pass
a pointer instead.

PASSING COMPLETION HANDLERS:

The pre-existing code passed completion handlers using C++ move
semantics. This is not yet supported by Swift/C++ interop
(rdar://162361370) so instead we have a wrap it in a
RefCountable wrapper.

RESTRICTIONS ON SWIFT TYPE USE:

The types in Swift can be forward-declared anywhere, but to actually
use them or know their size, we must include <WebKit-Swift.h>.
That may ONLY be done from .cpp files rather than .h files, to avoid
the risk of circular dependencies. (The generation of WebKit-Swift.h
depends on Swift being able to read potentially all of WebKit's header
files).

This constraint explains some of the delicate dance we do where
references to the Swift types are actually stored in std::unique_ptr;
this means the size of those types needs only be known by .cpp files
not by .h files. It's a bit of a pain and requires a few more heap
allocations, but for now, this is not performance-critical.

WHAT IS IN THIS COMMIT:

This commit contains all the above, but also a few other things
that merit explanation.

* More overloads of callMemberFunction which accept these two
  changes:
  * Use of a RefCountable<CompletionHandler> rather than CompletionHandler
  * Passing an IPC::Connection* rather than IPC::Connection&
  (We may not need so many overloads).
* messages.py supports receivers outside namespaces since Swift
  code is not namespaced
* messages.py code to generate the WebThingyMessageForwarder
  and WebThingyWeakRef
* messages.py code to dispatch the handler function call to 'target'
  rather than 'this', which is the same thing unless Swift is in use,
  in which case it's the thing returned by WebThingyMessageForwarder::getTarget
* messages.py includes WebKit-Swift.h if the handler is in Swift code
  not C++